requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
bootstrap4==0.1.0
Brlapi==0.8.3
Brotli==1.1.0
bs4==0.0.2
bumblebee-status==2.2.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cloudpathlib==0.19.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
confection==0.1.5
crispy-bootstrap4==2024.1
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
cymem==2.0.8
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.4
django-crispy-forms==2.1
django-filter==24.3
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
duplicity==0.8.21
elevenlabs==1.9.0
email_validator==2.2.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fake-useragent==1.5.1
fastapi==0.112.1
fastapi-cli==0.0.5
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
Flask==3.0.0
Flask-Babel==2.0.0
Flask-Cors==3.0.8
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.5.0
future==0.18.2
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
i3ipc==2.2.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
keybert==0.8.5
keyring==23.5.0
langcodes==3.4.1
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
lxml_html_clean==0.1.1
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.2.0
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.5
murmurhash==1.0.10
mutagen==1.47.0
natsort==8.0.2
netifaces==0.11.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
nodeenv==1.8.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
paramiko==2.9.3
parse==1.20.1
pexpect==4.8.0
Pillow==9.5.0
powerline-status==2.7
preshed==3.0.9
proglog==0.1.10
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.22
pycryptodomex==3.21.0
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyee==11.1.0
pygit2==1.15.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyppeteer==2.0.0
pyquery==2.0.0
pyRFC3339==1.1
pyright==1.1.361
pyseoanalyzer==2024.4.21
pysrt==1.1.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
rake-nltk==1.0.6
regex==2024.5.15
reportlab==3.6.8
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentence-transformers==2.7.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
ssh-import-id==5.11
starlette==0.38.2
sympy==1.12
systemd-python==234
thinc==8.3.2
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
ttystatus==0.38
typer==0.12.4
types-PyYAML==6.0.12.20240808
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.30.6
uvloop==0.20.0
w3lib==2.1.2
wadllib==1.3.6
wasabi==1.1.3
watchfiles==0.23.0
weasel==0.4.1
websockets==13.0.1
Werkzeug==3.0.3
wrapt==1.16.0
wxPython==4.0.7
xdg==5
xkbgroup==0.2.0
xkit==0.0.0
yarl==1.9.4
yt-dlp==2024.10.7
zipp==1.0.0
